Restore systems, all tests passing
